requirements.txt
streamlit
opencv-python-headless
numpy
pillow
tensorflow-cpu
requests